+++ ErsatzTV.Scanner/Application/Jellyfin/Commands/SynchronizeJellyfinLibraryById.cs
@@ -1,0 +1,6 @@
+﻿using ErsatzTV.Core;
+
+namespace ErsatzTV.Scanner.Application.Jellyfin;
+
+public record SynchronizeJellyfinLibraryById
+    (int JellyfinLibraryId, bool ForceScan) : IRequest<Either<BaseError, string>>;